Update Document Generator
