+++ src/components/Dashboard/Home/index.js
@@ -169,6 +169,8 @@ const db=getFirestore()
           onChange={(e) => setSearchTerm(e.target.value)}
         />
         <button onClick={handleSearch}>Search</button>
+        <button onClick={history.push("/pdf")}>Create RFI</button>
+        <button onClick={() => console.log('Button clicked')}>Create RFA</button>
       <SubNav currentFolder="root folder" />
       {adminFolders && adminFolders.length > 0 && (
         <>
